+++ requirements_all.txt
@@ -2480,7 +2480,7 @@ yeelight==0.7.10
 yeelightsunflower==0.0.10
 
 # homeassistant.components.yolink
-yolink-api==0.0.3
+yolink-api==0.0.4
 
 # homeassistant.components.youless
 youless-api==0.16
